Move all the projects controllers into /controllers/projects
